Decision-Agents/requirements.txt
google-adk>=1.22.1
google-genai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
redis>=4.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0